misc/HtmlTestSuite/results/test-0020.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Href - script element tests on markup</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="testcommon.js"/>
</head><body>
<div id="log"/>
<svg xmlns="http://www.w3.org/2000/svg" height="100" id="svg" viewbox="0 0 100 100" width="100">
  <script xmlns:xlink="http://www.w3.org/1999/xlink" href="testScripts/externalScript2.js" id="script"/>
</svg>
<script>
'use strict';

// Use an independent test file for markup testing because it may affect other
// tests.

test(function(t) {
  var script = document.getElementById('script');
  assert_equals(script.href.baseVal, 'testScripts/externalScript1.js');
  assert_equals(loadedScript(), 'externalScript1');
}, 'Test for loading external script by markup when setting both href and ' +
   'xlink:href');

</script>
</body>
</html>
